+++ backend/src/main/java/com/jarom/funbankapp/dto/DepositRequest.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 @Data
 public class DepositRequest {
-    private int accountId;
+    private Long accountId;
     private BigDecimal amount;
     private String description;
 }
